feat(TDD cycle: Refactor stage): Return a 'Defensive Copy'
Returns a 'Defensive Copy' instead of a 'Direct Reference' to the
internal list 'moments'.

Return a 'Direct Reference' to the internal list 'moments' can lead to
'unintended side effects' & 'compromise the INTEGRITY of the object's
state'. This violates the principle of 'encapsulation'.

Returning a 'Defensive Copy' provides complete independence. The
returned list can be modified by the caller without affecting the
internal list object. This ensure that the object's internal state
remains under its own control, adhering to the principle of
'encapsulation'.