feat(view): Implement 'MomentView' for handling user I/O.
This class hahdles 'the moment details' entered by the user (i.e. title,
description, emotion, and date) after prompting the user via the
terminal for the corresponding information.

It then verifies that the emotion exists & that the date is in valid
format.

Finally, it 'returns a string array' with all these details & sends the
user a message via the terminal indicating that the moment was added
successfully.