feat(TDD cycle: Green stage): Implement 'delete a single moment'.
It should contain the minimal amount of code needed to make the test
pass.

Implements the `deleteMoment` method to delete a single moment by ID.

This approach uses a `Stream` to remove an element from a list (it uses
a default method on the `Collection` interface, not the full "stream
API"). This is an efficient & powerful tool for processing data.